\name{workers}
\alias{startNWScluster}
\alias{startMPIcluster}
\alias{clusterSize}
\alias{clusterSize.nwscluster}
\alias{clusterSize.mpicluster}
\alias{closeCluster}
\alias{closeCluster.nwscluster}
\alias{closeCluster.mpicluster}
\alias{sinkWorkerOutput}
\title{Worker Management Functions}
\description{
The \code{startMPIcluster} function is used to start an MPI cluster.
The \code{startNWScluster} function is used to start an NWS cluster
for use with doMPI.
}
\keyword{utilities}
